refactor diet entry screen to use entry component
