+++ screens/DietEntryScreen.js
@@ -1,16 +1,21 @@
 import { Alert } from 'react-native'
 import React from 'react'
 import { useNavigation } from '@react-navigation/native';
+import { useState } from 'react'
 
 import { writeToDB } from '../firebase/FirebaseHelper';
-import DietEntry from '../components/DietEntry';
+import Entry from '../components/Entry';
 
 const DietEntryScreen = () => {
 
   const validCalories = /^[1-9]\d*$/; // positive integers without leading zeros
   const navigation = useNavigation();
+  const [description, setDescription] = useState('');
+  const [date, setDate] = useState(new Date()); // current date
+  const [dateText, setDateText] = useState('');
+  const [calories, setCalories] = useState('');
 
-  function handleSave(description, calories, date, dateText) {
+  function handleSave() {
     if (description && validCalories.test(calories) && dateText && date) {
       writeToDB({description, calories, date}, 'dietItems');
       navigation.goBack();
@@ -20,8 +25,29 @@ const DietEntryScreen = () => {
     }
   }
 
+  function handleChangeDate(selectedDate) {
+    setDate(selectedDate);
+    setDateText(selectedDate.toDateString());
+  }
+
+  function handleDescriptionChange(selectedDescription) {
+    setDescription(selectedDescription);
+  }
+
+  function handleCaloriesChange(selectedCalories) {
+    setCalories(selectedCalories);
+  }
+
   return (
-    <DietEntry 
+    <Entry 
+      type={'diet'}
+      title={description}
+      amount={calories}
+      date={date}
+      dateText={dateText}
+      changeTitleHandler={handleDescriptionChange}
+      changeAmountHandler={handleCaloriesChange}
+      changeDateHandler={handleChangeDate}
       saveHandler={handleSave}
     />
   )
